hwp_diff/requirements.txt
PySide6>=6.4.0
python-docx>=0.8.11
openpyxl>=3.1.0
scikit-learn>=1.0.0
lxml>=4.9.0
python-Levenshtein>=0.12.0
olefile>=0.47
